Move `getAppAddrWithProtocol` to `services/tshd/app.ts`
